move json api to a flask.Blueprint
