+++ fava/templates/editor.html
@@ -11,8 +11,8 @@
                 <option{% if source_file == file_path %} selected="selected"{% endif %}>{{ source_file }}</option>
             {% endfor %}
         </select>
-        <button id="source-editor-submit" type="submit" data-url="{{ url_for('api_source') }}">{{ _('Save') }}</button>
-        <button id="source-editor-format" type="submit" data-url="{{ url_for('api_format_source') }}">{{ _('Format') }}</button>
+        <button id="source-editor-submit" type="submit" data-url="{{ url_for('json_api.source') }}">{{ _('Save') }}</button>
+        <button id="source-editor-format" type="submit" data-url="{{ url_for('json_api.format_source') }}">{{ _('Format') }}</button>
     </div>
     <div id="source-editor-wrapper" class="source-editor-wrapper">
         <textarea id="source-editor" name="source" autofocus>{{ api.file.get_source(file_path) or '' }}</textarea>
