+++ fava/templates/overlays/_documents_upload.html
@@ -13,7 +13,7 @@
                 </select>
             </div>
             <div class="fieldset">
-                <button type="submit" id="document-upload-submit" data-url="{{ url_for('api_add_document') }}">Upload</button>
+                <button type="submit" id="document-upload-submit" data-url="{{ url_for('json_api.add_document') }}">Upload</button>
             </div>
         </form>
     </div>
