+++ fava/templates/overlays/_transaction_form.html
@@ -1,7 +1,7 @@
 <div id="transaction" class="transaction-overlay overlay-wrapper">
     <div class="overlay">
         <a href="#" class="close-overlay">(close)</a>
-        <form id="transaction-form" class="transaction-form" action="{{ url_for('api_add_transaction') }}">
+        <form id="transaction-form" class="transaction-form" action="{{ url_for('json_api.add_transaction') }}">
             <h3>New transaction:</h3>
             <div class="fieldset">
                 <label for="date">Date:</label>
@@ -15,8 +15,8 @@
             <div class="postings">
             </div>
             <div class="fieldset">
-                <button type="submit" id="transaction-form-submit" data-url="{{ url_for('api_add_document') }}">Save</button>
-                <button type="submit" class="muted" id="transaction-form-submit-and-new" data-url="{{ url_for('api_add_document') }}">Save and add new</button>
+                <button type="submit" id="transaction-form-submit" data-url="{{ url_for('json_api.add_document') }}">Save</button>
+                <button type="submit" class="muted" id="transaction-form-submit-and-new" data-url="{{ url_for('json_api.add_document') }}">Save and add new</button>
             </div>
         </form>
     </div>
